Spitted filter builder tests
